Implement single-task dumping
