+++ cmd/dump/Cargo.toml
@@ -20,6 +20,5 @@ lzss.workspace = true
 num-traits.workspace = true
 parse_int.workspace = true
 probe-rs.workspace = true
-rand.workspace = true
 zerocopy.workspace = true
 zip.workspace = true
