+++ cmd/dump/src/lib.rs
@@ -65,29 +65,25 @@
 //! ```
 //!
 
-use anyhow::{anyhow, bail, Context, Result};
+use anyhow::{bail, Result};
 use clap::{ArgGroup, CommandFactory, Parser};
 use core::mem::size_of;
 use hif::*;
 use humility::arch::ARMRegister;
 use humility::cli::Subcommand;
-use humility::core::{Core, NetAgent};
+use humility::core::Core;
+use humility::dump::{
+    task_areas, AgentCore, DumpAgent, DumpAgentExt, DumpArea, UdpDumpAgent,
+};
 use humility::hubris::*;
 use humility_cmd::hiffy::*;
 use humility_cmd::idol::{self, HubrisIdol};
 use humility_cmd::{Archive, Attach, Command, CommandKind, Validate};
-use humpty::*;
-use indexmap::IndexMap;
-use indicatif::{HumanBytes, HumanDuration};
-use indicatif::{ProgressBar, ProgressStyle};
+use humpty::{DumpAreaHeader, DumpSegment, DumpSegmentHeader, DumpTask};
+use indicatif::{HumanBytes, HumanDuration, ProgressBar, ProgressStyle};
 use num_traits::FromPrimitive;
-use rand::Rng;
 use std::cell::RefCell;
-use std::collections::BTreeMap;
-use std::collections::HashMap;
-use std::path::Path;
 use std::time::{Duration, Instant};
-use zerocopy::FromBytes;
 
 #[derive(Parser, Debug)]
 #[clap(
@@ -170,433 +166,6 @@ struct DumpArgs {
     list: bool,
 
     dumpfile: Option<String>,
-}
-
-////////////////////////////////////////////////////////////////////////////////
-
-//
-// When using the dump agent, we create our own ersatz Core
-//
-struct AgentCore {
-    flash: HubrisFlashMap,
-    ram_regions: BTreeMap<u32, Vec<u8>>,
-    registers: HashMap<ARMRegister, u32>,
-}
-
-impl AgentCore {
-    fn new(hubris: &HubrisArchive) -> Result<AgentCore> {
-        Ok(Self {
-            flash: HubrisFlashMap::new(hubris)?,
-            ram_regions: Default::default(),
-            registers: Default::default(),
-        })
-    }
-
-    fn add_ram_region(&mut self, addr: u32, contents: Vec<u8>) {
-        self.ram_regions.insert(addr, contents);
-    }
-
-    fn add_register(&mut self, reg: ARMRegister, val: u32) {
-        self.registers.insert(reg, val);
-    }
-
-    fn read_flash(&self, addr: u32, data: &mut [u8]) -> Result<()> {
-        if let Some((&base, &(size, offset))) =
-            self.flash.regions.range(..=addr).rev().next()
-        {
-            if base > addr || base + size <= addr {
-                //
-                // It's not here, and we have already tried RAM...
-                //
-                bail!("address 0x{:08x} not found", addr);
-            }
-
-            let start = (addr - base) as usize;
-            let roffs = offset + start;
-
-            if start + data.len() <= size as usize {
-                //
-                // This flash region wholly contains our desired region; copy
-                // it and leave.
-                //
-                data.copy_from_slice(
-                    &self.flash.contents[roffs..roffs + data.len()],
-                );
-
-                return Ok(());
-            }
-
-            let len = (size as usize) - start;
-            data[..len]
-                .copy_from_slice(&self.flash.contents[roffs..roffs + len]);
-
-            self.read_flash(addr + len as u32, &mut data[len..])
-        } else {
-            bail!("address 0x{:08x} not found", addr);
-        }
-    }
-
-    fn read(&mut self, addr: u32, data: &mut [u8]) -> Result<()> {
-        if let Some((&base, contents)) =
-            self.ram_regions.range(..=addr).rev().next()
-        {
-            if base > addr || base + (contents.len() as u32) <= addr {
-                //
-                // We don't have this in RAM -- pull it out of flash.
-                //
-                return self.read_flash(addr, data);
-            }
-
-            let start = (addr - base) as usize;
-
-            if start + data.len() <= contents.len() {
-                //
-                // This region -- and only this region -- contains our RAM.
-                // Copy it and leave.
-                //
-                data.copy_from_slice(&contents[start..start + data.len()]);
-                return Ok(());
-            }
-
-            //
-            // This region contains our RAM, but there is more.  Copy the bit
-            // that we want and recurse.
-            //
-            let len = contents.len() - start;
-            data[..len].copy_from_slice(&contents[start..contents.len()]);
-            self.read(addr + len as u32, &mut data[len..])
-        } else {
-            self.read_flash(addr, data)
-        }
-    }
-}
-
-impl Core for AgentCore {
-    fn info(&self) -> (String, Option<String>) {
-        panic!("unexpected call to AgentCore info");
-    }
-
-    fn read_8(&mut self, addr: u32, data: &mut [u8]) -> Result<()> {
-        self.read(addr, data)
-    }
-
-    fn read_reg(&mut self, reg: ARMRegister) -> Result<u32> {
-        match self.registers.get(&reg) {
-            Some(val) => Ok(*val),
-            None => bail!("unexpected read from register {reg:?}"),
-        }
-    }
-
-    fn write_reg(&mut self, reg: ARMRegister, _value: u32) -> Result<()> {
-        bail!("cannot write register {} over dump agent", reg);
-    }
-
-    fn write_word_32(&mut self, _addr: u32, _data: u32) -> Result<()> {
-        bail!("cannot write a word over dump agent");
-    }
-
-    fn write_8(&mut self, _addr: u32, _data: &[u8]) -> Result<()> {
-        bail!("cannot write a byte over dump agent");
-    }
-
-    fn halt(&mut self) -> Result<()> {
-        bail!("unexpected call to halt");
-    }
-
-    fn run(&mut self) -> Result<()> {
-        bail!("unexpected call to run");
-    }
-
-    fn step(&mut self) -> Result<()> {
-        bail!("can't step over dump agent");
-    }
-
-    fn init_swv(&mut self) -> Result<()> {
-        bail!("cannot enable SWV over dump agent");
-    }
-
-    fn read_swv(&mut self) -> Result<Vec<u8>> {
-        bail!("cannot read SWV over dump agent");
-    }
-
-    fn load(&mut self, _path: &Path) -> Result<()> {
-        bail!("cannot load flash over dump agent");
-    }
-
-    fn reset(&mut self) -> Result<()> {
-        bail!("cannot reset over dump agent");
-    }
-
-    fn reset_and_halt(&mut self, _dur: std::time::Duration) -> Result<()> {
-        bail!("cannot reset over dump agent");
-    }
-
-    fn wait_for_halt(&mut self, _dur: std::time::Duration) -> Result<()> {
-        bail!("cannot wait for halt over dump agent");
-    }
-}
-
-////////////////////////////////////////////////////////////////////////////////
-
-/// Trait which abstracts away dump agent interfaces
-trait DumpAgent {
-    fn core(&mut self) -> &mut dyn Core;
-
-    /// Initializes the dump regions
-    fn initialize_dump(&mut self) -> Result<()>;
-
-    /// Initializes a set of segments defined as `(address, length)` tuples
-    fn initialize_segments(&mut self, segments: &[(u32, u32)]) -> Result<()>;
-
-    /// Kicks off a dump
-    fn take_dump(&mut self) -> Result<()>;
-
-    /// General-purpose dump reading abstraction
-    ///
-    /// This function is given an iterator over areas, which are represented as
-    /// `index, offset` tuples.  It reads from those areas, chunking internally
-    /// as needed for efficiency.
-    ///
-    /// In the implementation, reading should terminate under three conditions:
-    /// - the iterator runs out
-    /// - the `cont` callback returns `false` (in which case the data for which
-    ///   it returned `false` is *not* returned)
-    /// - the read operation returns an `InvalidArea` error
-    ///
-    /// The latter is a little subtle, so it merits further explanation.  In
-    /// some cases, we may optimistically read dump areas that _may not exist_,
-    /// so we rely on this error to tell us when we've gone off the edge of the
-    /// map.  As an example, this occurs when using `--dump-agent-status` to
-    /// read every single area.
-    ///
-    /// The `cont` callback can also be used as a progress tracker.  Each time
-    /// it is called, it takes an incremental number of bytes that have been
-    /// read in the most recent operation.
-    #[allow(clippy::type_complexity)]
-    fn read_generic(
-        &mut self,
-        areas: &mut dyn Iterator<Item = (u8, u32)>,
-        cont: &mut dyn FnMut(u8, u32, &[u8]) -> Result<bool>,
-    ) -> Result<Vec<Vec<u8>>>;
-}
-
-/// This is the trait which does the vast majority of dumping
-///
-/// It's expected to use `read_generic` from a `DumpAgent`
-trait DumpAgentExt {
-    /// Helper function to make calling `read_generic` more pleasant
-    fn read_generic<I, F>(&mut self, areas: I, cont: F) -> Result<Vec<Vec<u8>>>
-    where
-        I: Iterator<Item = (u8, u32)>,
-        F: FnMut(u8, u32, &[u8]) -> Result<bool>;
-
-    ////////////////////////////////////////////////////////////////////////
-    // Everything beyond this point is implemented in terms of `read_generic`
-
-    /// Reads the root header
-    fn read_dump_header(&mut self) -> Result<DumpAreaHeader> {
-        self.read_dump_header_at(0)
-    }
-
-    /// Reads the header from the given dump area
-    fn read_dump_header_at(&mut self, i: u8) -> Result<DumpAreaHeader> {
-        let val = self.read_dump_area_start(i)?;
-        let (header, _task) = parse_dump_header(i as usize, &val)?;
-        Ok(header)
-    }
-
-    /// Reads 256 bytes from the start of the given dump area
-    ///
-    /// This is enough data to extract the header
-    fn read_dump_area_start(&mut self, i: u8) -> Result<Vec<u8>> {
-        // Read the header of this dump area
-        let mut val =
-            self.read_generic(std::iter::once((i, 0)), |_, _, _| Ok(true))?;
-        assert_eq!(val.len(), 1);
-        Ok(val.pop().unwrap())
-    }
-
-    /// Reads back a single dump area by index
-    fn read_dump_area(
-        &mut self,
-        index: u8,
-        progress: &mut dyn FnMut(usize),
-    ) -> Result<(DumpAreaHeader, Vec<u8>)> {
-        let val = self.read_dump_area_start(index)?;
-        let header = DumpAreaHeader::read_from_prefix(val.as_slice())
-            .ok_or_else(|| anyhow!("failed to read parse header"))?;
-
-        if header.magic != humpty::DUMP_MAGIC {
-            bail!("bad magic at dump offset {:#x}: {:x?}", index, header);
-        }
-
-        let size = size_of::<DumpAreaHeader>();
-
-        let mut out = vec![];
-        if header.written > size as u32 {
-            let max = val.len().min(header.written as usize);
-            let data = val[size..max].to_owned();
-            progress(data.len());
-            out.push(data);
-        }
-
-        let chunk_size = val.len();
-        assert_eq!(chunk_size, 256);
-
-        let mut chunks = vec![];
-        let mut offset = chunk_size.min(header.written as usize);
-        while offset < header.written as usize {
-            let len = chunk_size.min(header.written as usize - offset);
-            chunks.push(offset.try_into().unwrap());
-            offset += len;
-        }
-
-        // Read all of the region, one 256-byte chunk at a time
-        out.extend(self.read_generic(
-            chunks.into_iter().map(|offset| (index, offset)),
-            |_index, _offset, result| {
-                progress(result.len());
-                Ok(true)
-            },
-        )?);
-
-        // Collect chunks into a single vec
-        let mut out_data: Vec<u8> = out.into_iter().flatten().collect();
-
-        // Because we read in chunks, we may have extra data at the end here
-        out_data.resize(offset, 0u8);
-
-        Ok((header, out_data))
-    }
-
-    /// Reads dump headers from the target
-    ///
-    /// If `raw` is set, includes headers configured as `DUMPER_NONE`;
-    /// otherwise, stops once the first `DUMPER_NONE` header is encountered.
-    fn read_dump_headers(
-        &mut self,
-        raw: bool,
-    ) -> Result<Vec<(DumpAreaHeader, Option<DumpTask>)>> {
-        //
-        // Iterate over dump area headers.  Once we've hit an area that hasn't
-        // been dumped to, or an invalid area (denoting that all areas are
-        // full), we're done...
-        //
-        let results = self.read_generic(
-            (0..).map(|index| (index, 0)),
-            |index, _offset, val| {
-                let (header, _task) = parse_dump_header(index as usize, val)?;
-                if !raw && header.dumper == humpty::DUMPER_NONE {
-                    Ok(false)
-                } else {
-                    Ok(true)
-                }
-            },
-        )?;
-
-        results
-            .into_iter()
-            .enumerate()
-            .map(|(i, r)| parse_dump_header(i, &r))
-            .collect::<Result<Vec<_>>>()
-    }
-
-    fn read_dump(
-        &mut self,
-        area: Option<DumpArea>,
-        out: &mut AgentCore,
-    ) -> Result<Option<DumpTask>> {
-        let (base, headers, task) = {
-            // Read dump headers until the first empty header (DUMPER_NONE)
-            let all = self.read_dump_headers(false)?;
-
-            let area = match area {
-                None => None,
-                Some(DumpArea::ByIndex(ndx)) => Some(ndx),
-                Some(DumpArea::ByAddress(address)) => all
-                    .iter()
-                    .enumerate()
-                    .filter(|&(_, (header, _))| header.address == address)
-                    .map(|(ndx, _)| ndx)
-                    .next(),
-            };
-
-            match area {
-                None | Some(0) if all[0].1.is_none() => (
-                    0usize,
-                    all.iter()
-                        .map(|(h, _t)| *h)
-                        .collect::<Vec<DumpAreaHeader>>(),
-                    None,
-                ),
-
-                Some(ndx) => {
-                    let areas = task_areas(&all);
-                    match areas.get(&ndx) {
-                        None => {
-                            bail!("area {ndx} is invalid (--list to list)");
-                        }
-                        Some((task, headers)) => {
-                            (ndx, headers.clone(), Some(*task))
-                        }
-                    }
-                }
-
-                _ => {
-                    bail!("area must be explicitly specified (--list to list)");
-                }
-            }
-        };
-
-        let total = headers.iter().fold(0, |sum, header| sum + header.written);
-
-        let started = Instant::now();
-        let bar = ProgressBar::new(total as u64);
-        bar.set_style(
-            ProgressStyle::default_bar()
-                .template("humility: pulling [{bar:30}] {bytes}/{total_bytes}"),
-        );
-
-        let mut count = 0;
-        let mut contents = vec![];
-        for (ndx, _) in headers.iter().enumerate() {
-            let index = (ndx + base).try_into().unwrap();
-            let (_header, data) = self.read_dump_area(index, &mut |size| {
-                count += size;
-                bar.set_position(count as u64);
-            })?;
-            contents.extend(data.into_iter());
-        }
-
-        bar.finish_and_clear();
-
-        humility::msg!(
-            "pulled {} in {}",
-            HumanBytes(total as u64),
-            HumanDuration(started.elapsed())
-        );
-
-        //
-        // We have a dump!  On to processing...
-        //
-        process_dump(&headers[0], &contents, out, task)?;
-
-        Ok(task)
-    }
-}
-
-impl<'a> DumpAgentExt for Box<dyn DumpAgent + 'a> {
-    fn read_generic<I, F>(
-        &mut self,
-        mut areas: I,
-        mut cont: F,
-    ) -> Result<Vec<Vec<u8>>>
-    where
-        I: Iterator<Item = (u8, u32)>,
-        F: FnMut(u8, u32, &[u8]) -> Result<bool>,
-    {
-        DumpAgent::read_generic(self.as_mut(), &mut areas, &mut cont)
-    }
 }
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -834,141 +403,6 @@ impl<'a> DumpAgent for HiffyDumpAgent<'a> {
     }
 }
 
-////////////////////////////////////////////////////////////////////////////////
-
-struct UdpDumpAgent<'a> {
-    core: &'a mut dyn Core,
-}
-
-impl<'a> UdpDumpAgent<'a> {
-    /// Sends a remote dump command over the network
-    fn dump_remote_action(
-        &mut self,
-        msg: udp::Request,
-    ) -> Result<Result<udp::Response, udp::Error>> {
-        use udp::{Header, RequestMessage, ResponseMessage};
-        let mut rng = rand::thread_rng();
-        let header =
-            Header { version: udp::version::CURRENT, message_id: rng.gen() };
-        let mut buf = vec![
-            0u8;
-            std::cmp::max(
-                std::mem::size_of::<RequestMessage>(),
-                std::mem::size_of::<ResponseMessage>()
-            )
-        ];
-        let size = hubpack::serialize(&mut buf, &(header, msg))
-            .context("failed to serialize message")?;
-
-        // Send the packet out
-        self.core
-            .send(&buf[..size], NetAgent::DumpAgent)
-            .context("failed to send packet")?;
-
-        // Try to receive a reply
-        let size = self
-            .core
-            .recv(buf.as_mut_slice(), NetAgent::DumpAgent)
-            .context("failed to receive packet")?;
-
-        let (reply_header, rest): (udp::Header, _) =
-            hubpack::deserialize(&buf[..size])
-                .map_err(|_| anyhow!("deserialization of header failed"))?;
-        if reply_header.message_id != header.message_id {
-            bail!(
-                "message ID mismatch: {} != {}",
-                reply_header.message_id,
-                header.message_id
-            );
-        } else if reply_header.version < udp::version::MIN {
-            bail!(
-                "received reply with invalid version: {} < our min ({})",
-                reply_header.version,
-                udp::version::MIN
-            );
-        }
-
-        let (reply, _) = hubpack::deserialize(rest).map_err(|_| {
-            if reply_header.version > udp::version::CURRENT {
-                anyhow!(
-                    "deserialization of body failed, \
-                        version {} > our version {}",
-                    reply_header.version,
-                    udp::version::CURRENT
-                )
-            } else {
-                anyhow!(
-                    "deserialization of body failed, despite versions matching"
-                )
-            }
-        })?;
-
-        Ok(reply)
-    }
-}
-
-impl<'a> DumpAgent for UdpDumpAgent<'a> {
-    fn read_generic(
-        &mut self,
-        areas: &mut dyn Iterator<Item = (u8, u32)>,
-        cont: &mut dyn FnMut(u8, u32, &[u8]) -> Result<bool>,
-    ) -> Result<Vec<Vec<u8>>> {
-        let mut out = vec![];
-        for (index, offset) in areas {
-            let r = self
-                .dump_remote_action(udp::Request::ReadDump { index, offset })?;
-            match r {
-                Ok(udp::Response::ReadDump(d)) => {
-                    if !cont(index, offset, &d)? {
-                        break;
-                    } else {
-                        out.push(d.to_vec())
-                    }
-                }
-                Err(udp::Error::InvalidArea) => break,
-                _ => bail!("invalid reply: {r:?}"),
-            }
-        }
-        Ok(out)
-    }
-
-    fn core(&mut self) -> &mut dyn Core {
-        self.core
-    }
-
-    fn initialize_dump(&mut self) -> Result<()> {
-        let r = self.dump_remote_action(udp::Request::InitializeDump)?;
-        match r {
-            Ok(udp::Response::InitializeDump) => Ok(()),
-            _ => bail!("invalid response: {r:?}"),
-        }
-    }
-
-    fn initialize_segments(&mut self, segments: &[(u32, u32)]) -> Result<()> {
-        for &(address, length) in segments {
-            let r = self.dump_remote_action(udp::Request::AddDumpSegment {
-                address,
-                length,
-            })?;
-            match r {
-                Ok(udp::Response::AddDumpSegment) => (),
-                _ => bail!("invalid response: {r:?}"),
-            }
-        }
-        Ok(())
-    }
-
-    fn take_dump(&mut self) -> Result<()> {
-        let r = self.dump_remote_action(udp::Request::TakeDump)?;
-        match r {
-            Ok(udp::Response::TakeDump) => Ok(()),
-            _ => bail!("invalid response: {r:?}"),
-        }
-    }
-}
-
-////////////////////////////////////////////////////////////////////////////////
-
 fn emulate_dump(
     core: &mut dyn Core,
     task: Option<DumpTask>,
@@ -1074,186 +508,6 @@ fn emulate_task_dump_prep(
     Ok(area.address)
 }
 
-fn parse_dump_header(
-    area: usize,
-    buf: &[u8],
-) -> Result<(DumpAreaHeader, Option<DumpTask>)> {
-    let header = DumpAreaHeader::read_from_prefix(buf)
-        .ok_or_else(|| anyhow!("failed to parse dump area {area}"))?;
-
-    if header.magic != humpty::DUMP_MAGIC {
-        bail!("bad magic at in dump area {area}: {header:x?}");
-    }
-
-    //
-    // This is a little sleazy (or maybe even a lot sleazy?):  we know that if
-    // we have task dump here, the number of segments is sufficiently low to
-    // assure that we will have also slurped our task information -- and we
-    // know that our task information will immediately follow our segment
-    // headers.  (We take some of the sting off of this by at least checking
-    // the sizes when we read the dump headers -- but we still rely on the
-    // fact that any task information is dumped first.)  Gone fishin'...
-    //
-    let size = size_of::<DumpAreaHeader>();
-    let toffs =
-        size + header.nsegments as usize * size_of::<DumpSegmentHeader>();
-
-    let task = if toffs < buf.len() {
-        if let Some(DumpSegment::Task(t)) = DumpSegment::from(&buf[toffs..]) {
-            Some(t)
-        } else {
-            None
-        }
-    } else {
-        None
-    };
-
-    Ok((header, task))
-}
-
-//
-// Because single task dumps can spread over adjacent areas, they can be a
-// little tedious to process; iterate over all dump areas and return a map of
-// area indices to a task/vector of headers tuple.
-//
-fn task_areas(
-    headers: &[(DumpAreaHeader, Option<DumpTask>)],
-) -> IndexMap<usize, (DumpTask, Vec<DumpAreaHeader>)> {
-    let mut rval = IndexMap::new();
-
-    if headers[0].1.is_none() {
-        return rval;
-    }
-
-    let mut current = 0;
-    let mut areas = vec![headers[0].0];
-
-    for (ndx, (header, task)) in headers.iter().enumerate().skip(1) {
-        match task {
-            None => {
-                //
-                // If the header indicates that it's been written, it's
-                // a continuation of our previous header -- otherwise, we're
-                // done processing entirely.
-                //
-                if header.dumper != humpty::DUMPER_NONE {
-                    areas.push(*header);
-                } else {
-                    break;
-                }
-            }
-            Some(_task) => {
-                //
-                // We have hit a new task dump, so record what we have seen
-                // for our current task dump, and start a new one.
-                //
-                rval.insert(current, (headers[current].1.unwrap(), areas));
-                current = ndx;
-                areas = vec![*header];
-            }
-        }
-    }
-
-    rval.insert(current, (headers[current].1.unwrap(), areas));
-    rval
-}
-
-fn process_dump(
-    header: &DumpAreaHeader,
-    dump: &Vec<u8>,
-    agent: &mut AgentCore,
-    task: Option<DumpTask>,
-) -> Result<()> {
-    let nsegments = header.nsegments;
-    let mut offset = nsegments as usize * size_of::<DumpSegmentHeader>();
-
-    if offset > dump.len() {
-        bail!("in situ dump is short; missing {nsegments} segments");
-    }
-
-    if offset == dump.len() {
-        bail!("in situ dump is empty");
-    }
-
-    while offset < dump.len() {
-        let segment = match DumpSegment::from(&dump[offset..]) {
-            Some(segment) => segment,
-            None => {
-                bail!("short read at offset {offset}");
-            }
-        };
-
-        match segment {
-            DumpSegment::Task(t) => {
-                match task {
-                    None => {
-                        bail!("found unexpected task {t:?}");
-                    }
-                    Some(task) if task != t => {
-                        bail!("task mismatch: found {t:?}, expected {task:?}");
-                    }
-                    _ => {}
-                }
-
-                offset += size_of::<DumpTask>();
-                continue;
-            }
-
-            DumpSegment::Register(reg) => {
-                //
-                // These are register values; slurp them and continue.
-                //
-                if let Some(register) = ARMRegister::from_u16(reg.register) {
-                    agent.add_register(register, reg.value);
-                } else {
-                    let r = reg.register;
-                    bail!("unrecognized register {r:#x} at offset {offset}");
-                }
-
-                offset += size_of::<DumpRegister>();
-                continue;
-            }
-
-            DumpSegment::Data(data) => {
-                offset += size_of::<DumpSegmentData>();
-
-                let len = data.uncompressed_length as usize;
-
-                let mut contents = vec![0; len];
-                let limit = offset + data.compressed_length as usize;
-
-                humpty::DumpLzss::decompress(
-                    lzss::SliceReader::new(&dump[offset..limit]),
-                    lzss::SliceWriter::new(&mut contents),
-                )?;
-
-                agent.add_ram_region(data.address, contents[0..len].to_vec());
-                offset = limit;
-
-                while offset < dump.len()
-                    && dump[offset] == humpty::DUMP_SEGMENT_PAD
-                {
-                    offset += 1;
-                }
-            }
-
-            DumpSegment::Unknown(signature) => {
-                bail!(
-                    "unrecognized data with signature \
-                    {signature:x?} at offset {offset}"
-                );
-            }
-        }
-    }
-
-    Ok(())
-}
-
-enum DumpArea {
-    ByIndex(usize),
-    ByAddress(u32),
-}
-
 fn get_dump_agent<'a>(
     hubris: &'a HubrisArchive,
     core: &'a mut dyn Core,
@@ -1280,7 +534,7 @@ fn get_dump_agent<'a>(
             .unwrap_or(false)
     {
         humility::msg!("using UDP dump agent");
-        Ok(Box::new(UdpDumpAgent { core }))
+        Ok(Box::new(UdpDumpAgent::new(core)))
     } else {
         humility::msg!("using hiffy dump agent");
         Ok(Box::new(HiffyDumpAgent::new(hubris, core, subargs.timeout)?))
@@ -1292,7 +546,7 @@ fn dump_via_agent(
     core: &mut dyn Core,
     subargs: &DumpArgs,
 ) -> Result<()> {
-    let mut out = AgentCore::new(hubris)?;
+    let mut out = AgentCore::new(HubrisFlashMap::new(hubris)?);
     let started = Some(Instant::now());
     let mut area = subargs.area.map(DumpArea::ByIndex);
 
@@ -1518,7 +772,7 @@ fn dump_via_agent(
         //
         // If we're here, we have a dump in situ -- time to pull it.
         //
-        task = agent.read_dump(area, &mut out)?;
+        task = agent.read_dump(area, &mut out, true)?;
 
         //
         // If this was a whole-system dump, we will leave our state initialized
@@ -1572,8 +826,13 @@ fn dump_list(
         println!(
             "{area:>4} {:21} {:<10} {size}",
             match hubris.lookup_module(HubrisTask::Task(task.id.into())) {
-                Ok(module) => &module.name,
-                _ => "<unknown>",
+                Ok(module) => match headers[0].contents {
+                    humpty::DUMP_CONTENTS_SINGLETASK => module.name.to_owned(),
+                    humpty::DUMP_CONTENTS_TASKREGION =>
+                        format!("{} [region]", module.name.to_owned()),
+                    c => bail!("unknown contents type: {c}"),
+                },
+                _ => "<unknown>".to_owned(),
             },
             task.time,
         );
